+++ src/Gui/sp-gui.html/sp-gui-angular/@sp/ui-elements/src/lib/ui-elements.module.ts
@@ -1,32 +1,34 @@
-import { ToolbarDbComponent } from "./toolbar-db/toolbar-db.component";
-import { MatIconModule } from "@angular/material/icon";
+import { DragDropModule } from "@angular/cdk/drag-drop";
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { FormsModule } from "@angular/forms";
 import { MatButtonModule } from "@angular/material/button";
-import { MatToolbarModule } from "@angular/material/toolbar";
 import { MatButtonToggleModule } from "@angular/material/button-toggle";
 import { MatCardModule } from "@angular/material/card";
+import { MatDialogModule } from "@angular/material/dialog";
 import { MatExpansionModule } from "@angular/material/expansion";
-import { MatGridListModule } from "@angular/material/grid-list";
 import { MatFormFieldModule } from "@angular/material/form-field";
-import { CommonModule } from "@angular/common";
-import { SpToolbarButtonComponent } from "./sp-toolbar-button/sp-toolbar-button.component";
-import { SpSolutionDescComponent } from "./sp-solution-desc/sp-solution-desc.component";
-import { ProblemInfoComponent } from "./problem-info/problem-info.component";
-import { ToolbarEditComponent } from "./toolbar-edit/toolbar-edit.component";
-import { ToolbarPieceComponent } from "./toolbar-piece/toolbar-piece.component";
-import { NgModule } from "@angular/core";
-import { QuillModule } from "ngx-quill";
-import { FormsModule } from "@angular/forms";
+import { MatGridListModule } from "@angular/material/grid-list";
+import { MatIconModule } from "@angular/material/icon";
 import { MatInputModule } from "@angular/material/input";
 import { MatSelectModule } from "@angular/material/select";
-import { DragDropModule } from "@angular/cdk/drag-drop";
-import { QuillInitializeService } from "./services/quillInitialize.service";
-import { ProblemDefinitionsComponent } from "./problem-definitions/problem-definitions.component";
-import { ProblemAuthorsComponent } from "./problem-authors/problem-authors.component";
-import { ProblemPublicationComponent } from "./problem-publication/problem-publication.component";
+import { MatToolbarModule } from "@angular/material/toolbar";
 import { MatTooltipModule } from "@angular/material/tooltip";
+import { QuillModule } from "ngx-quill";
+import { ConfirmDialogComponent } from './confirm-dialog/confirm-dialog.component';
 import { FileExplorerComponent } from "./file-explorer/file-explorer.component";
-import { ToolbarEngineComponent } from "./toolbar-engine/toolbar-engine.component";
 import { FileSourceSelectorComponent } from "./file-source-selector/file-source-selector.component";
+import { ProblemAuthorsComponent } from "./problem-authors/problem-authors.component";
+import { ProblemDefinitionsComponent } from "./problem-definitions/problem-definitions.component";
+import { ProblemInfoComponent } from "./problem-info/problem-info.component";
+import { ProblemPublicationComponent } from "./problem-publication/problem-publication.component";
+import { QuillInitializeService } from "./services/quillInitialize.service";
+import { SpSolutionDescComponent } from "./sp-solution-desc/sp-solution-desc.component";
+import { SpToolbarButtonComponent } from "./sp-toolbar-button/sp-toolbar-button.component";
+import { ToolbarDbComponent } from "./toolbar-db/toolbar-db.component";
+import { ToolbarEditComponent } from "./toolbar-edit/toolbar-edit.component";
+import { ToolbarEngineComponent } from "./toolbar-engine/toolbar-engine.component";
+import { ToolbarPieceComponent } from "./toolbar-piece/toolbar-piece.component";
 
 @NgModule({
   declarations: [
@@ -42,6 +44,7 @@ import { FileSourceSelectorComponent } from "./file-source-selector/file-source-
     ProblemPublicationComponent,
     FileExplorerComponent,
     FileSourceSelectorComponent,
+    ConfirmDialogComponent,
   ],
   imports: [
     CommonModule,
@@ -59,6 +62,7 @@ import { FileSourceSelectorComponent } from "./file-source-selector/file-source-
     MatInputModule,
     MatSelectModule,
     DragDropModule,
+    MatDialogModule,
   ],
   exports: [
     ToolbarDbComponent,