+++ src/Gui/sp-gui.html/sp-gui-angular/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+- [x] [#119](https://github.com/dardino/scacchi-painter/issues/119) Added functionality "Add" and "Remove" to the database
 - [x] [#106](https://github.com/dardino/scacchi-painter/issues/106) Added commands to add a problem into current db
 - [x] [#103](https://github.com/dardino/scacchi-painter/issues/103) Problem selection from database page
 - [x] Update to latest Angular 15